Conflicts:
	azure-mgmt-sql/src/main/java/com/microsoft/azure/management/sql/SqlServer.java
	azure-mgmt-sql/src/main/java/com/microsoft/azure/management/sql/implementation/SqlServerManager.java

+++ gulpfile.js
@@ -12,9 +12,170 @@ var mappings = {
         'args': '-FT 1',
         'modeler': 'CompositeSwagger'
     },
+    'eventhub': {
+        'dir': 'azure-mgmt-eventhub',
+        'source': 'arm-eventhub/2015-08-01/swagger/EventHub.json',
+        'package': 'com.microsoft.azure.management.eventhub',
+        'args': '-FT 1'
+    },
+    'servicefabric': {
+        'dir': 'azure-mgmt-servicefabric',
+        'source': 'arm-servicefabric/2016-09-01/swagger/servicefabric.json',
+        'package': 'com.microsoft.azure.management.servicefabric',
+        'args': '-FT 1'
+    },
+    'notificationhubs': {
+        'dir': 'azure-mgmt-notificationhubs',
+        'source': 'arm-notificationhubs/2017-04-01/swagger/notificationhubs.json',
+        'package': 'com.microsoft.azure.management.notificationhubs',
+        'args': '-FT 1'
+    },
+    'analysisservices': {
+        'dir': 'azure-mgmt-analysisservices',
+        'source': 'arm-analysisservices/2016-05-16/swagger/analysisservices.json',
+        'package': 'com.microsoft.azure.management.analysisservices',
+        'args': '-FT 1'
+    },
+    'automation': {
+        'dir': 'azure-mgmt-automation',
+        'source': 'arm-automation/compositeAutomation.json',
+        'package': 'com.microsoft.azure.management.authorization',
+        'args': '-FT 1',
+        'modeler': 'CompositeSwagger'
+    },
+    'billing': {
+        'dir': 'azure-mgmt-billing',
+        'source': 'arm-billing/2017-04-24-preview/swagger/billing.json',
+        'package': 'com.microsoft.azure.management.billing',
+        'args': '-FT 1'
+    },
+    'cognitiveservices': {
+        'dir': 'azure-mgmt-cognitiveservices',
+        'source': 'arm-cognitiveservices/2017-04-18/swagger/cognitiveservices.json',
+        'package': 'com.microsoft.azure.management.cognitiveservices',
+        'args': '-FT 1'
+    },
+    'consumption': {
+        'dir': 'azure-mgmt-consumption',
+        'source': 'arm-consumption/2017-04-24-preview/swagger/consumption.json',
+        'package': 'com.microsoft.azure.management.consumption',
+        'args': '-FT 1'
+    },
+    'customerinsights': {
+        'dir': 'azure-mgmt-customerinsights',
+        'source': 'arm-customer-insights/2017-04-26/swagger/customer-insights.json',
+        'package': 'com.microsoft.azure.management.customerinsights',
+        'args': '-FT 1'
+    },
+    'devtestlab': {
+        'dir': 'azure-mgmt-devtestlab',
+        'source': 'arm-devtestlabs/2016-05-15/swagger/DTL.json',
+        'package': 'com.microsoft.azure.management.devtestlab',
+        'args': '-FT 1'
+    },
+    'insights': {
+        'dir': 'azure-mgmt-insights',
+        'source': 'arm-insights/compositeInsightsManagementClient.json',
+        'package': 'com.microsoft.azure.management.gallery',
+        'args': '-FT 1',
+        'modeler': 'CompositeSwagger'
+    },
+    'intune': {
+        'dir': 'azure-mgmt-intune',
+        'source': 'arm-intune/2015-01-14-preview/swagger/intune.json',
+        'package': 'com.microsoft.azure.management.intune',
+        'args': '-FT 1'
+    },
+    'iothub': {
+        'dir': 'azure-mgmt-devices',
+        'source': 'arm-iothub/2017-01-19/swagger/iothub.json',
+        'package': 'com.microsoft.azure.management.devices',
+        'args': '-FT 1'
+    },
+    'logic': {
+        'dir': 'azure-mgmt-logic',
+        'source': 'arm-logic/2016-06-01/swagger/logic.json',
+        'package': 'com.microsoft.azure.management.logic',
+        'args': '-FT 1'
+    },
+    'machinelearning': {
+        'dir': 'azure-mgmt-machinelearning',
+        'source': 'arm-machinelearning/2017-01-01/swagger/webservices.json',
+        'package': 'com.microsoft.azure.management.machinelearning',
+        'args': '-FT 1'
+    },
+    'media': {
+        'dir': 'azure-mgmt-media',
+        'source': 'arm-mediaservices/2015-10-01/swagger/media.json',
+        'package': 'com.microsoft.azure.management.media',
+        'args': '-FT 1'
+    },
+    'operationalinsights': {
+        'dir': 'azure-mgmt-operationalinsights',
+        'source': 'arm-operationalinsights/compositeOperationalInsights.json',
+        'package': 'com.microsoft.azure.management.operationalinsights',
+        'args': '-FT 1',
+        'modeler': 'CompositeSwagger'
+    },
+    'powerbi': {
+        'dir': 'azure-mgmt-powerbi',
+        'source': 'arm-powerbiembedded/2016-01-29/swagger/powerbiembedded.json',
+        'package': 'com.microsoft.azure.management.powerbi',
+        'args': '-FT 1'
+    },
+    'recoveryservices': {
+        'dir': 'azure-mgmt-recoveryservices',
+        'source': 'arm-recoveryservices/compositeRecoveryServicesClient.json',
+        'package': 'com.microsoft.azure.management.recoveryservices',
+        'args': '-FT 1',
+        'modeler': 'CompositeSwagger'
+    },
+    'recoveryservicesbackup': {
+        'dir': 'azure-mgmt-recoveryservicesbackup',
+        'source': 'arm-recoveryservicesbackup/2016-12-01/swagger/backupManagement.json',
+        'package': 'com.microsoft.azure.management.recoveryservicesbackup',
+        'args': '-FT 1'
+    },
+    'recoveryservicessiterecovery': {
+        'dir': 'azure-mgmt-recoveryservicessiterecovery',
+        'source': 'arm-recoveryservicessiterecovery/2016-08-10/swagger/service.json',
+        'package': 'com.microsoft.azure.management.recoveryservicessiterecovery',
+        'args': '-FT 1'
+    },
+    'relay': {
+        'dir': 'azure-mgmt-relay',
+        'source': 'arm-relay/2017-04-01/swagger/relay.json',
+        'package': 'com.microsoft.azure.management.relay',
+        'args': '-FT 1'
+    },
+    'servermanagement': {
+        'dir': 'azure-mgmt-servermanagement',
+        'source': 'arm-servermanagement/2016-07-01-preview/swagger/servermanagement.json',
+        'package': 'com.microsoft.azure.management.servermanagement',
+        'args': '-FT 1'
+    },
+    'storsimple8000series': {
+        'dir': 'azure-mgmt-storsimple8000series',
+        'source': 'arm-storsimple8000series/2017-06-01/swagger/storsimple.json',
+        'package': 'com.microsoft.azure.management.storsimple8000series',
+        'args': '-FT 1'
+    },
+    'streamanalytics': {
+        'dir': 'azure-mgmt-streamanalytics',
+        'source': 'arm-streamanalytics/compositeStreamAnalytics.json',
+        'package': 'com.microsoft.azure.management.streamanalytics',
+        'args': '-FT 1',
+        'modeler': 'CompositeSwagger'
+    },
     'graphrbac': {
         'dir': 'azure-mgmt-graph-rbac',
         'source': 'arm-graphrbac/1.6/swagger/graphrbac.json',
+        'package': 'com.microsoft.azure.management.graphrbac',
+        'args': '-FT 1'
+    },
+    'authorization': {
+        'dir': 'azure-mgmt-graph-rbac',
+        'source': 'arm-authorization/2015-07-01/swagger/authorization.json',
         'package': 'com.microsoft.azure.management.graphrbac',
         'args': '-FT 1'
     },
@@ -64,12 +225,6 @@ var mappings = {
         'args': '-FT 1',
         'modeler': 'CompositeSwagger'
     },
-    'graph.rbac': {
-        'dir': 'azure-mgmt-graph-rbac',
-        'source': 'arm-graphrbac/compositeGraphRbacManagementClient.json',
-        'package': 'com.microsoft.azure.management.graph.rbac',
-        'args': '-FT 1'
-    },
     'redis': {
         'dir': 'azure-mgmt-redis',
         'source': 'arm-redis/2016-04-01/swagger/redis.json',
@@ -78,13 +233,13 @@ var mappings = {
     },
     'search': {
         'dir': 'azure-mgmt-search',
-        'source': 'arm-search/2015-02-28/swagger/search.json',
+        'source': 'arm-search/2015-08-19/swagger/search.json',
         'package': 'com.microsoft.azure.management.search',
         'args': '-FT 1'
     },
     'trafficmanager': {
         'dir': 'azure-mgmt-trafficmanager',
-        'source': 'arm-trafficmanager/2015-11-01/swagger/trafficmanager.json',
+        'source': 'arm-trafficmanager/2017-05-01/swagger/trafficmanager.json',
         'package': 'com.microsoft.azure.management.trafficmanager',
         'args': '-FT 1'
     },
@@ -128,7 +283,7 @@ var mappings = {
     },
     'batch': {
         'dir': 'azure-mgmt-batch',
-        'source': 'arm-batch/2015-12-01/swagger/BatchManagement.json',
+        'source': 'arm-batch/2017-05-01/swagger/BatchManagement.json',
         'package': 'com.microsoft.azure.management.batch',
         'args': '-FT 1'
     },
@@ -150,6 +305,44 @@ var mappings = {
         'source': 'arm-dns/2016-04-01/swagger/dns.json',
         'package': 'com.microsoft.azure.management.dns',
         'args': '-FT 1'
+    },
+    'servicebus': {
+        'dir': 'azure-mgmt-servicebus',
+        'source': 'arm-servicebus/2015-08-01/swagger/servicebus.json',
+        'package': 'com.microsoft.azure.management.servicebus',
+        'args': '-FT 1'
+    },
+    'monitor': {
+        'dir': 'azure-mgmt-monitor',
+        'source': 'arm-monitor/compositeMonitorManagementClient.json',
+        'package': 'com.microsoft.azure.management.monitor',
+        'args': '-FT 1',
+        'modeler': 'CompositeSwagger'
+    },
+    'monitor-dataplane': {
+        'dir': 'azure-mgmt-monitor',
+        'source': 'monitor/compositeMonitorClient.json',
+        'package': 'com.microsoft.azure.management.monitor',
+        'args': '-FT 1 -ServiceName Monitor',
+        'modeler': 'CompositeSwagger'
+    },
+    'containerregistry': {
+        'dir': 'azure-mgmt-containerregistry',
+        'source': 'arm-containerregistry/2017-03-01/swagger/containerregistry.json',
+        'package': 'com.microsoft.azure.management.containerregistry',
+        'args': '-FT 1',
+    },
+    'scheduler': {
+        'dir': 'azure-mgmt-scheduler',
+        'source': 'arm-scheduler/2016-03-01/swagger/scheduler.json',
+        'package': 'com.microsoft.azure.management.scheduler',
+        'args': '-FT 1'
+    },
+    'cosmosdb': {
+        'dir': 'azure-mgmt-cosmosdb',
+        'source': 'arm-documentdb/2015-04-08/swagger/documentdb.json',
+        'package': 'com.microsoft.azure.management.cosmosdb',
+        'args': '-FT 1',
     }
 };
 
@@ -161,13 +354,13 @@ gulp.task('default', function() {
     Object.keys(mappings).forEach(function(i) {
         console.log('\t' + i.magenta);
     });
-    console.log("--autorest\n\tThe version of AutoRest. E.g. 0.15.0, or the location of AutoRest repo, E.g. E:\\repo\\autorest");
+    console.log("--autorest\n\tThe version of AutoRest. E.g. 1.0.1-20170222-2300-nightly, or the location of AutoRest repo, E.g. E:\\repo\\autorest");
     console.log("--autorest-args\n\tPasses additional argument to AutoRest generator");
 });
 
 var specRoot = args['spec-root'] || "https://raw.githubusercontent.com/Azure/azure-rest-api-specs/master";
 var projects = args['projects'];
-var autoRestVersion = '1.0.0-Nightly20170209'; // default
+var autoRestVersion = 'latest'; // default
 if (args['autorest'] !== undefined) {
     autoRestVersion = args['autorest'];
 }
@@ -175,21 +368,10 @@ var autoRestArgs = args['autorest-args'];
 var autoRestExe;
 
 gulp.task('codegen', function(cb) {
-    var nugetSource = 'https://www.myget.org/F/autorest/api/v2';
-    if (autoRestVersion.match(/[0-9]+\.[0-9]+\.[0-9]+.*/)) {
-        autoRestExe = 'packages\\autorest.' + autoRestVersion + '\\tools\\AutoRest.exe';
-        nugetExe = "tools\\nuget.exe";
-        if (process.platform !== 'win32') {
-            nugetExe = "mono tools/nuget.exe";
-            autoRestExe = autoRestExe.replace(/\\/g, "/");
-            exec('chmod +x ' + autoRestExe);
-            autoRestExe = "mono " + autoRestExe;
-        }
-        exec(nugetExe + ' install AutoRest -Source ' + nugetSource + ' -Version ' + autoRestVersion + ' -o packages', function(err, stdout, stderr) {
-            console.log(stdout);
-            console.error(stderr);
-            handleInput(projects, cb);
-        });
+    if (autoRestVersion.match(/[0-9]+\.[0-9]+\.[0-9]+.*/) ||
+        autoRestVersion == 'latest') {
+        autoRestExe = 'autorest ---version=' + autoRestVersion;
+        handleInput(projects, cb);
     } else {
         autoRestExe = autoRestVersion + "/src/core/AutoRest/bin/Debug/netcoreapp1.0/AutoRest.dll";
         autoRestExe = "dotnet " + autoRestExe;
@@ -217,7 +399,9 @@ var handleInput = function(projects, cb) {
 
 var codegen = function(project, cb) {
     var outputDir = mappings[project].dir + '/src/main/java/' + mappings[project].package.replace(/\./g, '/');
-    deleteFolderRecursive(outputDir);
+    if (!args['preserve']) {
+        deleteFolderRecursive(outputDir);
+    }
     console.log('Generating "' + project + '" from spec file ' + specRoot + '/' + mappings[project].source);
     var generator = 'Azure.Java.Fluent';
     if (mappings[project].fluent !== null && mappings[project].fluent === false) {
@@ -231,8 +415,10 @@ var codegen = function(project, cb) {
                         ' -CodeGenerator ' + generator +
                         ' -Namespace ' + mappings[project].package +
                         ' -Input ' + specRoot + '/' + mappings[project].source +
-                        ' -outputDirectory ' + mappings[project].dir + '/src/main/java/' + mappings[project].package.replace(/\./g, '/') +
+                        ' -outputDirectory ' + outputDir +
                         ' -Header MICROSOFT_MIT_NO_CODEGEN' +
+                        ' -skipValidation true' +
+                        ' -RegenerateManager true ' +
                         ' -' + autoRestArgs;
     if (mappings[project].args !== undefined) {
         cmd = cmd + ' ' + mappings[project].args;
